Network configuration (refs #1416)
- dns configuration dialog
- validation of array objects

+++ tests/TKMON/Tests/Model/System/InterfacesTest.php
@@ -18,7 +18,9 @@ class InterfacesTest extends \PHPUnit_Framework_TestCase
 
     public static function tearDownAfterClass()
     {
-        unlink(self::$interfaceFile);
+        if (file_exists(self::$interfaceFile)) {
+            unlink(self::$interfaceFile);
+        }
     }
 
     public function testStep1()
